through sub_0207D8EC

+++ include/unk_0208805C.h
@@ -1,7 +1,7 @@
 #ifndef POKEHEARTGOLD_UNK_0208805C_H
 #define POKEHEARTGOLD_UNK_0208805C_H
 
-int RatioToInt(int hp, int maxHp, int a2);
+u8 RatioToInt(int hp, int maxHp, int a2);
 int sub_020880B0(u16 hp, u16 maxHp, int a2);
 int sub_020880CC(int a0, HeapID heapId);
